chore: address issue #159
